Jest Testing
- API endpoints test with Jest

+++ catalogService/prisma/migrations/20231105171704_init/migration.sql
@@ -4,11 +4,11 @@ CREATE TABLE `Car` (
     `model` VARCHAR(191) NOT NULL,
     `make` VARCHAR(191) NOT NULL,
     `type` VARCHAR(191) NOT NULL,
-    `price` DECIMAL(65, 30) NOT NULL,
+    `price` INTEGER NOT NULL,
     `img` VARCHAR(191) NOT NULL,
     `year` INTEGER NOT NULL,
     `history` VARCHAR(191) NOT NULL,
-    `milage` DECIMAL(65, 30) NOT NULL,
+    `milage` INTEGER NOT NULL,
 
     PRIMARY KEY (`id`)
 ) DEFAULT CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci;
